Reorg Coproduct instances
